Admin: make the fleet map the home screen — semantic zoom, a warped time axis, and a two-tab shell
Eight rounds of design work on the admin, done outside the pipeline because
it is a matter of taste rather than a specification.

The map becomes the root route. It draws every chain on one canvas with
semantic zoom: chains at overview, cards at mid, and the full stage pipeline
(work -> audit as a decision gate -> rework loop) when zoomed into an item.
History runs to the left on a warped time axis that clips dead gaps and packs
settled chains into a shared strip, so 180 days of work stays legible without
the past swamping the present.

Around it: an attention rail for items that need the operator, suggestion
ghosts that promote to work items, release containers, a (+) affordance that
adds a dependent without touching layout, camera bounds that stop the view
drifting off the content, and a ruler that reads out the instant under the
pointer (or says plainly that it is inside a cut).

The shell drops the sidebar for two tabs — Map view and Queue view — with the
remaining pages folded into a grouped menu.

The layout takes no camera input, so zooming never moves a node: covered by
LayoutInvarianceTests across the whole zoom range.

+++ tools/CodeyBox.Admin/src/CodeyBox.Admin.Web/Models/QuotaDto.cs
@@ -24,7 +24,10 @@ public sealed class QuotaProbeDto
 
 public sealed class QuotaSnapshotDto
 {
-    public int AvailablePct { get; set; }
+    /// <summary>Percentage of quota remaining. The orchestrator returns a decimal
+    /// (an agent can sit at 0.416%), so this must not be an int — deserialising it
+    /// as one threw and took the whole fleet map down with it.</summary>
+    public double AvailablePct { get; set; }
     public bool? IsKnown { get; set; }
     public DateTimeOffset? ResetAt { get; set; }
     public string? Notes { get; set; }
